#118 rename: MSW notifications folder 정리
